+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Changed
 - Makes verifiers per shard configurable
+- Total token calculation per shard
 
 ## [0.1.0] - 2018-10-03
 ### Added